refactor: 레이아웃 구조 변경건 적용
