+++ app/layout.module.scss
@@ -24,22 +24,30 @@
   align-items: center;
   overflow-y: auto;
   overflow-x: hidden;
-  padding-top: rem(90px);
+  padding-top: rem(30px);
+
+  @media (max-width: $mantine-breakpoint-xs) {
+    padding-top: rem(15px);
+  }
 }
 
 .container {
+  display: flex;
+  flex-direction: column;
   max-width: 1080px;
   width: 100%;
   flex: 1 1 auto;
+  padding-left: rem(10px);
+  padding-right: rem(10px);
+  gap: rem(20px);
 
   @media (max-width: $mantine-breakpoint-sm) {
-    padding-top: rem(60px);
     padding-left: rem(15px);
     padding-right: rem(15px);
+    gap: rem(10px);
   }
 
   @media (max-width: $mantine-breakpoint-xs) {
-    padding-top: rem(30px);
     padding-left: rem(10px);
     padding-right: rem(10px);
   }
